requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amocrm-api==2.6.1
anyio==3.7.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Flask==3.1.0
Brotli==1.1.0
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
cryptography==43.0.3
frozenlist==1.5.0
greenlet==3.1.1
h11==0.12.0
h2==4.1.0
hpack==4.0.0
hstspreload==2024.10.1
httpcore==0.15.0
httpx==0.25.1
hyperframe==6.0.1
idna==2.10
kaitaistruct==0.10
multidict==6.1.0
outcome==1.3.0.post0
packaging==24.1
phonenumbers==9.0.3
pipdeptree==2.24.0
playwright==1.48.0
propcache==0.2.0
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pyee==12.0.0
PyJWT==2.10.1
pyOpenSSL==24.2.1
pyparsing==3.2.0
PySocks==1.7.1
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
PyVirtualDisplay==3.0
requests==2.32.3
rfc3986==1.5.0
selenium==4.25.0
selenium-wire==5.1.0
setuptools==75.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
telebot==0.0.5
trio==0.27.0
trio-websocket==0.11.1
TwoCaptcha==0.0.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.2.3
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
yarl==1.17.1
zstandard==0.23.0
